requirements.txt
moviepy==1.0.3
pillow==9.5.0

google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.2.1

imageio>=2.5
imageio-ffmpeg>=0.2.0
tqdm>=4.11.2
requests>=2.8.1
